Add fbx and bvh mobile share (#13)
Add fbx and bvh mobile share, The file is downloaded, except if the file
has already been downloaded, once the share button has been pressed.
This means there's a bit of load time when you want to share a file that
still needs to be downloaded. Currently, no spinner is displayed but
this will change in the future.

To make it reusable I've expanded the IconTextButton to be wrapped by
ShareFileButton.